+++ tests/test_trie.py
@@ -1,29 +1,17 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, unicode_literals
-import string
-import random
 import tempfile
 import pickle
 
 import pytest
 
 import marisa_trie
-
-def get_random_words(count):
-    russian = 'абвгдеёжзиклмнопрстуфхцчъыьэюя'
-    alphabet = russian + string.ascii_lowercase
-
-    def random_word(length):
-        return "".join([random.choice(alphabet) for x in range(random.randint(1,length))])
-
-    return list(set(random_word(10) for y in range(count)))
+from .utils import get_random_words
 
 
 def test_build():
-    trie = marisa_trie.Trie()
     keys = get_random_words(1000)
-
-    trie.build(keys)
+    trie = marisa_trie.Trie(keys)
 
     for key in keys:
         assert key in trie
@@ -36,13 +24,13 @@ def test_contains():
     trie = marisa_trie.Trie()
     assert 'foo' not in trie
 
-    trie.build(['foo'])
+    trie = marisa_trie.Trie(['foo'])
     assert 'foo' in trie
     assert 'f' not in trie
 
 def test_key_id():
     words = ['foo', 'bar', 'f']
-    trie = marisa_trie.Trie().build(words)
+    trie = marisa_trie.Trie(words)
     for word in words:
         key_id = trie.key_id(word)
         assert trie.restore_key(key_id) == word
@@ -61,7 +49,7 @@ def test_saveload():
     fd, fname = tempfile.mkstemp()
 
     words = ['foo', 'bar', 'f']
-    trie = marisa_trie.Trie().build(words)
+    trie = marisa_trie.Trie(words)
 
     with open(fname, 'w') as f:
         trie.write(f)
@@ -77,7 +65,7 @@ def test_saveload():
 def test_mmap():
     fd, fname = tempfile.mkstemp()
     words = get_random_words(1000)
-    trie = marisa_trie.Trie().build(words)
+    trie = marisa_trie.Trie(words)
     with open(fname, 'w') as f:
         trie.write(f)
 
@@ -89,11 +77,10 @@ def test_mmap():
 
 def test_dumps_loads():
     words = get_random_words(1000)
-    trie = marisa_trie.Trie().build(words)
-    data = trie.dumps()
+    trie = marisa_trie.Trie(words)
+    data = trie.tobytes()
 
-    trie2 = marisa_trie.Trie()
-    trie2.loads(data)
+    trie2 = marisa_trie.Trie().frombytes(data)
 
     for word in words:
         assert word in trie2
@@ -101,7 +88,7 @@ def test_dumps_loads():
 
 def test_pickling():
     words = get_random_words(1000)
-    trie = marisa_trie.Trie().build(words)
+    trie = marisa_trie.Trie(words)
 
     data = pickle.dumps(trie)
     trie2 = pickle.loads(data)
@@ -115,11 +102,11 @@ def test_len():
     trie = marisa_trie.Trie()
     assert len(trie) == 0
 
-    trie.build(['foo', 'f', 'bar'])
+    trie = marisa_trie.Trie(['foo', 'f', 'bar'])
     assert len(trie) == 3
 
 def test_prefixes():
-    trie = marisa_trie.Trie().build(['foo', 'f', 'foobar', 'bar'])
+    trie = marisa_trie.Trie(['foo', 'f', 'foobar', 'bar'])
     assert trie.prefixes('foobar') == ['f', 'foo', 'foobar']
     assert trie.prefixes('foo') == ['f', 'foo']
     assert trie.prefixes('bar') == ['bar']
@@ -129,11 +116,26 @@ def test_prefixes():
 
 def test_keys():
     keys = ['foo', 'f', 'foobar', 'bar']
-    trie = marisa_trie.Trie().build(keys)
+    trie = marisa_trie.Trie(keys)
     assert set(trie.keys()) == set(keys)
 
 def test_keys_prefix():
     keys = ['foo', 'f', 'foobar', 'bar']
-    trie = marisa_trie.Trie().build(keys)
+    trie = marisa_trie.Trie(keys)
     assert set(trie.keys('fo')) == set(['foo', 'foobar'])
     assert trie.keys('foobarz') == []
+
+
+#def test_int_trie():
+#    data = {'foo': 10, 'f': 20, 'foobar':30, 'bar': 40}
+#    trie = marisa_trie.IntTrie().build(data)
+#
+#    for key, value in data.items():
+#        assert trie[key] == value
+#
+#    trie['foo'] = 500
+#    assert trie['foo'] == 500
+#
+#    # trie is frozen, no new keys are allowed
+#    with pytest.raises(KeyError):
+#        trie['z'] = 200